+++ packages/eslint-plugin/src/rules/no-object-literal-type-assertion.ts
@@ -1,8 +1,3 @@
-/**
- * @fileoverview Forbids an object literal to appear in a type assertion expression
- * @author Armano <https://github.com/armano2>
- */
-
 import { AST_NODE_TYPES, TSESTree } from '@typescript-eslint/typescript-estree';
 import * as util from '../util';
 
@@ -14,7 +9,7 @@ type Options = [
 type MessageIds = 'unexpectedTypeAssertion';
 
 export default util.createRule<Options, MessageIds>({
-  name: 'no-object-literal-type-assertions',
+  name: 'no-object-literal-type-assertion',
   meta: {
     type: 'problem',
     docs: {