+++ packages/eslint-plugin/src/util/createRule.ts
@@ -26,7 +26,7 @@ type CreateRuleMeta<TMessageIds extends string> = {
 // This function will get much easier to call when this is merged https://github.com/Microsoft/TypeScript/pull/26349
 // TODO - when the above rule lands; add type checking for the context.report `data` property
 export function createRule<
-  TOptions extends Readonly<any[]>,
+  TOptions extends any[],
   TMessageIds extends string,
   TRuleListener extends RuleListener = RuleListener
 >({
@@ -48,7 +48,7 @@ export function createRule<
       ...meta,
       docs: {
         ...meta.docs,
-        url: `https://github.com/typescript-eslint/typescript-eslint/blob/${version}/packages/eslint-plugin/docs/rules/${name}.md`,
+        url: `https://github.com/typescript-eslint/typescript-eslint/blob/v${version}/packages/eslint-plugin/docs/rules/${name}.md`,
         extraDescription: meta.docs.tslintName
           ? [`\`${meta.docs.tslintName}\` from TSLint`]
           : undefined,