+++ packages/typescript-estree/src/tsconfig-parser.ts
@@ -1,5 +1,3 @@
-'use strict';
-
 import path from 'path';
 import ts from 'typescript';
 import { Extra } from './parser-options';
@@ -29,6 +27,8 @@ const knownWatchProgramMap = new Map<
  * There may be more than one per file if a file is shared between projects
  */
 const watchCallbackTrackingMap = new Map<string, ts.FileWatcherCallback>();
+
+const parsedFilesSeen = new Set<string>();
 
 /**
  * Holds information about the file currently being linted
@@ -73,7 +73,7 @@ export function calculateProjectParserOptions(
   // Update file version if necessary
   // TODO: only update when necessary, currently marks as changed on every lint
   const watchCallback = watchCallbackTrackingMap.get(filePath);
-  if (typeof watchCallback !== 'undefined') {
+  if (parsedFilesSeen.has(filePath) && typeof watchCallback !== 'undefined') {
     watchCallback(filePath, ts.FileWatcherEventKind.Changed);
   }
 
@@ -176,6 +176,7 @@ export function calculateProjectParserOptions(
     results.push(program);
   }
 
+  parsedFilesSeen.add(filePath);
   return results;
 }